backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.3
pydantic==2.5.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.23.8
